A first try at using LSP built into neovim
Didn't manage to it it to actually working. I got it to the point where
gopls starts up, but for some reason my config still isn't connecting
the .go filetype to that server, so none of the autocomplete, etc
starts.

It made me realise that I like would struggle to figure out the config
for the language server itself to make it play nice with a monorep so
I'll stop here.

Resources:
https://github.com/mrnugget/vimconfig/pull/1/files